requirements.txt
google-genai>=1.0.0
gTTS>=2.3.2
requests>=2.31.0
moviepy==1.0.3
Pillow>=10.0.0
numpy>=1.24.0
imageio>=2.25.0
imageio-ffmpeg>=0.4.7
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
python-dotenv>=1.0.0
pyyaml>=6.0
